requirements.txt
attrs==20.2.0
bleach==3.2.1
certifi==2020.6.20
chardet==3.0.4
colorama==0.4.4
dbutils-gonzalo123==0.1.0
docutils==0.16
idna==2.10
iniconfig==1.1.1
keyring==21.4.0
packaging==20.4
pkginfo==1.6.1
pluggy==0.13.1
psycopg2-binary==2.8.6
py==1.9.0
Pygments==2.7.2
pyparsing==2.4.7
pytest==6.1.2
pytest-env==0.6.2
readme-renderer==28.0
requests==2.24.0
requests-toolbelt==0.9.1
rfc3986==1.4.0
six==1.15.0
toml==0.10.1
tqdm==4.51.0
twine==3.2.0
urllib3==1.25.11
webencodings==0.5.1
